Workpiece documentation state (#52)
* preps for documentation state
- model validators extracted to validators/index.js
- model stubs for documentation of WorkpieceState.js
- documentation state
- added fiel types "collection" and "map"

* saving work on models
