+++ src/mobX/models/workpieces/documentation/DocLyricsModel.js
@@ -1,0 +1,3 @@
+import BaseModel, { FieldType, Field } from "../../../BaseModel"
+import { observable, action, computed } from "mobx"
+export default class DocLyricsModel extends BaseModel {}
