+++ src/mobX/models/workpieces/documentation/DocReleaseModel.js
@@ -1,0 +1,3 @@
+import BaseModel, { FieldType, Field } from "../../../BaseModel"
+import { observable, action, computed } from "mobx"
+export default class DocReleaseModel extends BaseModel {}
